+++ CHANGELOG.md
@@ -12,6 +12,7 @@ CentOS-6 6.10 x86_64 - Varnish Cache 4.1.
 - Updates source image to [1.9.0](https://github.com/jdeathe/centos-ssh/releases/tag/1.9.0).
 - Updates pattern for static assets to include web fonts and SVG images and remove SWF.
 - Removes response header that indicate Varnish version.
+- Adds `VARNISH_AUTOSTART_VARNISHD_WRAPPER` for disabling varnishd autostart.
 
 ### 1.4.4 - 2018-06-22